+++ CHANGELOG.md
@@ -4,6 +4,7 @@ SPDX-License-Identifier: CC0-1.0
 -->
 
 ## [NEXT] - Unreleased
+  * New API: new_from_app1_segment allows reading metadata from a buffer.
   * Require Rust 1.63 as the minimum supported language version.
   * Adopt 2021 edition of the language.
   * Dependency upgrades.